removing unnecessary import statements and refactoring page folder
