+++ src/components/Footer/Footer.js
@@ -4,9 +4,9 @@ import "../../App.scss";
 class Footer extends React.Component {
   render() {
     return (
-      <div className="footer bg-dark text-light">
+      <div className="bg-dark text-light">
         <hr />
-        {/* Copyright @ MangoDB 2020. All rights reserved. */}
+         Copyright @ MangoDB 2020. All rights reserved.
       </div>
     );
   }
