+++ src/components/UserProfile/components/Accomplishments/index.js
@@ -1,7 +1,4 @@
 import * as React from "react";
-import "../../../../App.css";
-import "../../../../App.scss";
-import "../../UserProfile.css";
 import Carousel from "react-bootstrap/Carousel";
 
 class Accomplishments extends React.Component {
