+++ src/components/UserProfile/components/Avatar/index.js
@@ -1,7 +1,4 @@
 import * as React from "react";
-import "../../../../App.css";
-import "../../../../App.scss";
-import "../../UserProfile.css";
 
 class Avatar extends React.Component {
