+++ src/scenes/Pages/FeedPage/index.js
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Feed from "../../components/Feed";
+import Feed from "../../../components/Feed";
 
 class FeedPage extends React.Component {
   render() {
